+++ crates/regui/src/backend/mod.rs
@@ -1,1 +1,4 @@
 pub(crate) mod shapes;
+
+#[cfg(feature = "wgpu")]
+pub(crate) mod texture;